provide 1st class support for zod schemas in useForm
useForm still accepts a well-supported AbstractSchema `schema` key.
However, the function is now properly overloaded to support zod as a 1st
class schema. This means you can now pass a zod schema directly to
useForm and get the same type safety (under the hood, we identify zod
schemas and generate corresponding abstract schemas). This was done as a
dev convenience