declare types for new input tracker concept
an InputTracker is essentially a "spy" that records all values that we
submit to setValue to be committed to the internal form store. The
purpose of this input tracker is to eventually enable an api like:

const userCount = getValue('count').input // ref
const formCount = getValue('count').output // ref

aka (my preferred syntax):

const { input, output } = getCount('count')

at any level of the tree (input/output are computed on demand)

This addresses the implicit design challenge of building a form that
either only stores correct values or allows the actual user submitted
values, at the expense of type safety and interim correctness. In the
input/output model, we thing that the value that is stored in the form
(i.e. 'output') is a function of the user's input:

output = f(input)

In this way, the internal form state always has a valid type signature,
which helps DX. However, we can always spy on the persistent user
submitted state via value.input

This is very much experimental and subject to change