Implement meta API for getValue() (#3)
* refactor where code is located

pnpm prepack still works as expected

* demo of a non-working example we want to work on

* declare types for new input tracker concept

an InputTracker is essentially a "spy" that records all values that we
submit to setValue to be committed to the internal form store. The
purpose of this input tracker is to eventually enable an api like:

const userCount = getValue('count').input // ref
const formCount = getValue('count').output // ref

aka (my preferred syntax):

const { input, output } = getCount('count')

at any level of the tree (input/output are computed on demand)

This addresses the implicit design challenge of building a form that
either only stores correct values or allows the actual user submitted
values, at the expense of type safety and interim correctness. In the
input/output model, we thing that the value that is stored in the form
(i.e. 'output') is a function of the user's input:

output = f(input)

In this way, the internal form state always has a valid type signature,
which helps DX. However, we can always spy on the persistent user
submitted state via value.input

This is very much experimental and subject to change

* implement function to flatten objects and return the leaf nodes as values

Related to this, I'm now thinking of making the API something like:

const count = getValue("count") // ref of count in form
const userSuppliedValue = count.spy // ref of user supplied value

this allows users to access the form state as normal, but also
introspect the user value if needed

* add introspection to setValue based on flattedObjectWithBaseKey function

* implement basic use-input-tracker-store and auto-register form trackers

* quick playground app demonstrating what is currently possible

the eslint/ts errors are expected, I'm intentionally forcing
unreasonable keys/values into the setValue function to see what happens.

* rename types file and loosen type restriction on InitialState

* improve DeepPartial type inference

never type was filtering out `unknown` concrete type, leading to bugs in
other places. This fixes that.

* extend types for getValue response using CurrentValueWithContext type

* implement functions for flattening and reconstructing obj from flat version

* add context api to getValue

* add meta tracker to setValue

* implement stable use-meta-tracker-store

* add meta tracker store logic to abstract form composable

* refactor the use-form composable to fix type inference for internal zod adapter

* wip playground showing what's pososible

yes, there are 2 eslint errors. it's fine, this is a test branch

* remove unnecessary ts ignore linting rule

* no longer publish automatically (use github actions page going forward)

That was a useful feature when I was trying to get things to work.