move away from attr transformation
I did a lot of experimental work in another throw-away repo and
concluded that using the created lifecycle hook to handle SSR-adjacent
logic is more reasonable. The goal is now to implement a v-supermodel
directive that takes in an object that wraps a ref, as a way of
achieving two-way binding.