refactor: update XModelValue type and update directives appropriately
This was done because the toRefs that is computed for innerRef is a
readonly ref by default. implementing a setter pattern seemed reasonable
enough at the time and got everything working easily. you could, ofc,
choose to make the toRef readable. refactor seems to be working, based
on very light testing