wip (stable): implement element store that can bind elements to listeners
This allows me to understand which set of elements are connected to a
given input and expose simple functions for the v-xmodel directives to
directly register/de-register elements. the last thing would be creating
a simple reactive state for developers to tap into the input state. I'll
implement this in a future commit, but right now the console.logs in the
browser show that we're correctly managing our events of choice during
the element's lifecycle, as well as deregistering the element on unmount