Implement SSR-safe v-xmodel, register function, and getElementState (#5)
* declare transform types on useForm configuration (for abstract and zod configurations)

* update attribute names

* wip: got the plugin auto-registratation working

* update pnpm-lock deps

* move away from attr transformation

I did a lot of experimental work in another throw-away repo and
concluded that using the created lifecycle hook to handle SSR-adjacent
logic is more reasonable. The goal is now to implement a v-supermodel
directive that takes in an object that wraps a ref, as a way of
achieving two-way binding.

* wip v-supermodel is now working but options flash

* wip: update deps

* wip: remove console log

* implement select-transform

This node transformer identifies select nodes in the application's AST
and processes their underlying option nodes, if any, searching
recursively until all appropriate options are selected. This makes the
select element on the frontend SSR-safe. Resonably tested, although no
tests exist. This transform is also a blueprint for making other
elements SSR safe.

Key idea: modify props prior to the compiler generating a codegenNode
for the vnode under consideration. `nodeTransforms` in the vue compiler
is generally a pretty safe place to do this (as demonstrated here).
directive node transforms have subtle complications with child vnodes
that will have precomputed codegenNodes, hence the decision not to
construct a directive node transform -- keeps things simpler.

* change plugin name and switch import

* wip[stable]: implement clientside xmodel directive

Works for the following input types:

1. text
2. checkbox (single, multiple)
3. radio

Also works for select elements (single, multiple) and textarea elements.

I'm not so sure how gracefully the implementation will carry over to
custom components (eg. iirc we check if an element is a select element
somewhere, which could break things -- will refactor later, as needed).

overall, native elements are decently supported at this point

we are also able to retrieve the html element references properly

* wip: update module.ts declaration

mostly externalizing vue dependencies

* wip: demo app in a random working state

* refactor: ready to delete the original directive

* wip: export type MetaTrackerValue

* fix: use args of function with _ rather than variables beyond the function scope

* feat: implement a transform for inputs of various types + text area element

expected to be compatible with xmodel, works well in light testing so
far.

* refactor: update XModelValue type and update directives appropriately

This was done because the toRefs that is computed for innerRef is a
readonly ref by default. implementing a setter pattern seemed reasonable
enough at the time and got everything working easily. you could, ofc,
choose to make the toRef readable. refactor seems to be working, based
on very light testing

* wip: implementing register function (incomplete but working)

* refactor: update how the register function is used

* feat: the playground with a random example demonstrating that register is working

* fix: minor improvement for type assertion within zod adapter

* wip (stable): implement element store that can bind elements to listeners

This allows me to understand which set of elements are connected to a
given input and expose simple functions for the v-xmodel directives to
directly register/de-register elements. the last thing would be creating
a simple reactive state for developers to tap into the input state. I'll
implement this in a future commit, but right now the console.logs in the
browser show that we're correctly managing our events of choice during
the element's lifecycle, as well as deregistering the element on unmount

* wip: progress on the register utility

I may move this function to wherever the setValueFactory is

* feat: allow v-xmodel to manage the deregistration of elements

* update use-abstract-form implementation with latest registerFactory

* checkpoint: current demo of what's possible

* update ImportMeta interface

* progress on the use-element-store composable (in a stable state)

* implement an element state api

* implement function caching in the register factory function

I was worried about potentially recomputing the same functions for large
forms (avoiding memory leaks, listener de-registration bugs, etc), This
is defensive code, I didn't actually notice any issues

* finally expose getElementState method from useAbstractForm

Consider this API highly experimental

* app.vue showing what's possible right now

* refactor: switched from currentValue to value

I'll implement currentValue when I run into a need for currentValue vs
previousValue (tbh using a watcher is the vue way to achieve this, so
we're fine)

* chore: remove unused comments