Implement getState API for detailed context about leaf nodes of a form (internal, meta, and DOM info)
* update dependencies

* declare FormSummaryValue type for tracking computed state based on internal state

* wip: improve register logic

The main change here is that we no longer attempt to guess the html
element states (focused, blurred, touched, etc) on the server, but
rather send a null placeholder. We also make sure that isConnected is
always in sync with the appropriate html element states

* chore: remove comment

* fix: in flatten-object, use startsWith for basePath search, not includes

* refactor: switch from ref call to computed function (chatgpt called this out in review)

* fix: take care of subtle bugs with how the meta tracker computes its next state

* feat: implement function that updates form summary values reliably, work into use-form-store

* refactor: improve use-element-store logic, remove redundant code, declare `touched` attribute here

* rename getElementState to getState

getState returns a more descriptive state for the input, typing in the
internal and DOM states, as well as convenient metadata like updateAt
and such.

also, simplify how we gain access to metaTracker

* app.vue showing what's possible right now