fix: remove external declarations from module.ts
I noticed that the library builds perfectly in a test consumer app if I
ONLY use the build config, hence this change. my current theory is that
the module.ts code tried to run in the consumer app, which would lead to
issues. the build config, on the other hand, would only run when
generating the module dist