Fix: Move build configuration to build.config.ts so that consumer app's build properly (#23)
* refactor: refactors @types d.ts files to normal types

* refactor: remove global type (#app -> nuxt/app)

I wasn't using nuxt/app before due to an issue with how I was running
the library build in module.ts -- this should be resolved now

* remove #app declaration in tsconfig.json

* chore: remove @vue/runtime-core dependency

I never noticed that over time, I ended up implementing the library
without explicitly relying on runtime-core imports. The vue library is
good enough

* feat: explicitly signaling that the lib is built in env without auto-imports

* fix: remove external declarations from module.ts

I noticed that the library builds perfectly in a test consumer app if I
ONLY use the build config, hence this change. my current theory is that
the module.ts code tried to run in the consumer app, which would lead to
issues. the build config, on the other hand, would only run when
generating the module dist

* feat: eslint declaration to force specific lodash-es import style

* chore: update various utils imports (due to changes with lodash-es, @types)

* chore: update lib/core/composables import types

* chore: update runtime/composables import type paths

* chore: update type import paths in runtime/adapters/zod/index.ts

* chore: remove process-dist.ts from tsconfig

* refactor: switch to autoimport=true because of playground app

* eslint added a comma

* fix: coerce import.meta.client to bool for type safety

* chore: update deps and revalidate lock file