fix: remove hardcoded dist specification in tsconfig.json
This tiny bug took over two weeks for me to understand/resolve 👀.
Essentially, unbuild tries to do some clever things to generate types
that line up with the src directory. the outDir and declarationDir
settings inferfere with this, so `nuxt-module-build build` fails to
generate the types (or attempts to put them in a nonexistent folder, not
sure what the case is). The give away of this is the smaller dist size.

All the types are now getting compiled into the dist folder. This PR was
focused on getting types to show up in dist. We will tackle actually
cleaning up types and such in future PR's. Thanks for attending my TED
talk, cheers!