chore: declare @chemical-x/forms/types alias in playground for type safety
Note: this ONLY exposes the types from the src folder, nothing else.
Thus, module logic is unaffected, only a type-inference improvement for
the playground/app.vue template (v-xmodel generally screams when
appropriate).

Note that v-xmodel doesn't scream if NOTHING is passed to it (example,
something like <input v-xmodel />) . This is strictly not correct, but I
don't have the bandwidth to implement an eslint config that gets
injected into the client app to handle this at the moment. Not the
biggest concern, but just wanted to acknowledge the ts shortfall here.
We may prioritize this bit of type-safety depending on practical usage.