feat(set-utilities): implement difference/intersection utilities for sets
Rationale: I wasn't aware that .difference is an ESNext thing at the
moment, so the library was breaking in some consumer applications that
weren't running with the latest ESlib. I wrote these utilities to
capture the essence of Set.difference and Set.intersection using ES6
style code. I didn't want to do a polyfill/shim because that's something
that would run in the user's env. and I wanna be respectful about that.
Utility functions feel cleaner in that regard. Hoping this resolves
things.