Bug: Ensure that difference/intersection operations work in older node environments (#73)
* chore: removed unused code

* feat(set-utilities): implement difference/intersection utilities for sets

Rationale: I wasn't aware that .difference is an ESNext thing at the
moment, so the library was breaking in some consumer applications that
weren't running with the latest ESlib. I wrote these utilities to
capture the essence of Set.difference and Set.intersection using ES6
style code. I didn't want to do a polyfill/shim because that's something
that would run in the user's env. and I wanna be respectful about that.
Utility functions feel cleaner in that regard. Hoping this resolves
things.

* chore(dev): stop using esnext, we need to be more aware of old eslib issues

I think we may be able to figure out how to get rollup to compile our
code in es6 (which I don't believe it is currently doing 👀) and then get
back to using sweet, sweet, ESNext. But yeah, given that this is a
consumer library where we can't control the environment in which the
@chemical-x/forms runs, this is probably the safest thing in general

* refactor: update useformstore to use setdifference and setintersection

This refactor attempts to satisfy older browser/node environments where
fancy methods like Set.difference and Set.intersection are not or can
not be supported for whatever reason.

* feat: add tsc checker script to package.json

This does a good job of catching eslint issues related to eslib.
Interestingly, going down to ES2020 only flagged the set method issues.

* chore: removed irrelevant test code

* refactor: bring back ./nuxt/tsconfig.json for sensible defaults

* refactor: do not manually specify tsconfig module resolution

Rely on nuxt module builder and ./nuxt/tsconfig.json to figure it out