fix(minor): add workflow step for publish-npm.yml to prepare dev environment (#77)
* fix(publish.yml): add pnpm dev:prepare step prior to test in publish.yml

* chore(publish.yml): rename workflow file to make its purpose more clear

* refactor: switch tsconfig lib from es2015 to es2020

Reason: we need APIs like import.meta etc, that rely on ES2020, so our
target build is ES2020. Thus, it probably doesn't make sense to build in
ES6. One could make the argument that working in an ES6 env expands the
scope of what can run, but we're losing out on a lot of ES2020 features.
The solution imo is to implement solid tests that target various node
environments (as is done already via .github/workflows/matrix.yml)