fix(tsconfig.json): remove extended .nuxt tsconfig (#80)
At build time, we couldn't reference the tsconfig.json file in the .nuxt
folder, which led to build errors. Simply removing the extended
reference (which I added in a recent previous PR thinking I was solving
something, before I understood the root cause) resolves this build
issue.

In addition, module is set to "ES2020" to line up with the lib and
build.config.ts build target