chore(eslint/prettier): format the codebase consistently
To format the codebase, it's `pnpm lint:fix --no-cache`. I did this
beforehand. Going forward, this generally shouldn't be necessary because
of the eslint and prettier extensions in VSCode helping keep the code
tidy and consistent as you type.

Note:

I'm not particularly opinionated about the way the code is written, as
long as it's highly readable and consistent. The rules I am using are
largely derived from me asking AI models what it thinks devs often
gravitate to. If you feel strongly about a certain eslint/prettier
decision, by all means create an issue for it, and the suggestion might
take precedence in the codebase. Thanks!