feat/chore: define eslint flat config + run eslint and prettier on repository (#86)
* try to commit with an an invalid title

* feat: add commitlint and husky

* does this commit get flagged?

* another try from the cli instead of lazygit

* add prepare script, remove analyze script

* refactor(githooks): update husky, commitlint, add @commitlint/config-conventional

* test(githooks): remove random test file

* refactor(husky): lint on commit, test/build on save

I will likely bring in lint-staged or something, this is just a test

* fix(husky): install husky dev dependency

* feat(readme.md): add nodejs test suite status badge to README.md

* feat(lint-staged): add lint-staged dev dep and pattern-matching keys

* refactor(husky): switch pre-commit hook to use lint-staged

* feat(husky): add tsc type checking step to pre-commit hook

I'll come back and see if I can run tsc on only the staged files

* refactor: include more folders that tsconfig should be aware of for eslint

* feat(vscode): add useful eslint/prettier settings to .vscode/settings.json

* chore: delete unnecessary files

* feat: bring in new eslint/prettier/nuxt plugins to improve formatting/linting

I'll survey the dependencies later, as some deps were abandoned along
the way.

* feat: declare .eslintignore, .prettierignore, and .prettierrc.cjs files

Fun fact, this wasn't supposed to include the .prettierrc.cjs file.
However, I was curious what would happen if I staged a file while
lazygit was processing a commit, and it added it without making sure it
was legit 👀. Just something to be aware of

* feat: declare basic eslint-plugin-nuxt module (no @types)

* feat(eslint): declare flat config that auto-generates globals for nuxt auto-imports

* chore(eslint/prettier): format the codebase consistently

To format the codebase, it's `pnpm lint:fix --no-cache`. I did this
beforehand. Going forward, this generally shouldn't be necessary because
of the eslint and prettier extensions in VSCode helping keep the code
tidy and consistent as you type.

Note:

I'm not particularly opinionated about the way the code is written, as
long as it's highly readable and consistent. The rules I am using are
largely derived from me asking AI models what it thinks devs often
gravitate to. If you feel strongly about a certain eslint/prettier
decision, by all means create an issue for it, and the suggestion might
take precedence in the codebase. Thanks!

* chore: regenerated pnpm-lock.yaml

I deleted the contents of pnpm-lock.yaml from github because merging it
with the file coming from the main branch involved a lot of changes.
This should be good, although in general, dependencies still need to be
reviewed, some are currently unused and will need to be removed.